sre-recipes/requirements.txt
click>=7.1.0
pyyaml>-5.4.1